Internal: build Stone separately from the SDK
Problem: the build time of the SDK reaches ~20 minutes, which is not
acceptable to users. This is due to the compilation of Stone in the
`build.rs` file.

Solution: we now build and install Stone from a Shell script that is
isolated from the Rust compilation process. Developers must now run this
script before attempting to run tests, but this accelerates the build
process significantly.